Small Changes see description
Added in heart containers, currently it triggers losing a heart by hitting spacebar(projectile). We can quickly end a Game End screen and options to restart or quit.

Fixed the projectile shooting in the correct direction when invoked, before it would only fire forward even if character is turned around

Logic to decerement heart is there right now, just call MainCharacter->lives-=1 I want to create a stun animation method in Main Character that decrements life and reacts to running into an enemy object or if an enemy shoots a projectile and it connects.
Right now doing this comparision of projectile to enemy or main character to enemy requires double for loop(parsing projectile-> parsing enemy) so I haven't implemented it because I think we can do something better. Double FOR is going to be laggy but we can do it as worst case.